rename to AbstractServiceException
